requirements.txt
# Core framework
flask==3.0.3
gunicorn==22.0.0
flask-sqlalchemy==3.1.1
flask-mail==0.10.0
flask-limiter==3.7.0

# AI / Vision
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.3.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# Task queue
celery==5.3.6
redis==5.0.4

# Auth / Security
werkzeug==3.0.3
cryptography==42.0.8

# Metrics
prometheus-client==0.20.0

# Templates
Jinja2==3.1.4

# HTTP
httpx==0.27.0
requests==2.32.3

# Config
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-flask==1.3.0
pytest-cov==5.0.0
faker==25.2.0
